backend/requirements.txt
django==5.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.3.1
plaid-python==15.1.0
python-dotenv==1.1.0 